se agregaron las validaciones y el cors

+++ backend/services/locality_service.py
@@ -1,9 +1,12 @@
 from extensions import db # Instancia de SQLAlchemy (maneja la DB)
 from models.locality_model import Locality
+from services.utils import normalize_data
+from exceptions import BusinessError
 
 #crea una localidad
 def create_locality(data):
-    locality = Locality(**data)
+    resp = normalize_data(data)
+    locality = Locality(**resp)
     db.session.add(locality) # Agrega a la sesión (pendiente de guardar)
     db.session.commit()  # Guarda en la base de datos
     return locality  # Devuelve el objeto creado
@@ -18,8 +21,9 @@ def get_all_localities():
 
 #Actualiza localidad
 def update_locality(locality,data):
+    resp = normalize_data(data)
      # Recorre el JSON recibido
-    for key,value in data.items():
+    for key,value in resp.items():
         setattr(locality,key,value)
         # setattr = asigna dinámicamente (ej: locality.name = "nuevo valor")
     db.session.commit() # Guarda cambios en la DB
@@ -27,5 +31,10 @@ def update_locality(locality,data):
 
 #Elimina localidad
 def delete_locality(locality):
+    if locality.contacts:
+        raise BusinessError(
+            "No se puede eliminar: tiene contactos asociados",
+            409
+        )
     db.session.delete(locality) # Marca para eliminar
     db.session.commit() # Ejecuta eliminación en la DB
